+++ sdk/dotnet/CloudWatch/EventPermission.cs
@@ -10,7 +10,9 @@ using Pulumi.Serialization;
 namespace Pulumi.Aws.CloudWatch
 {
     /// <summary>
-    /// Provides a resource to create a CloudWatch Events permission to support cross-account events in the current account default event bus.
+    /// Provides a resource to create an EventBridge permission to support cross-account events in the current account default event bus.
+    /// 
+    /// &gt; **Note:** EventBridge was formerly known as CloudWatch Events. The functionality is identical.
     /// 
     /// ## Example Usage
     /// ### Account Access