+++ sdk/dotnet/Sagemaker/NotebookInstance.cs
@@ -13,8 +13,7 @@ namespace Pulumi.Aws.Sagemaker
     /// Provides a Sagemaker Notebook Instance resource.
     /// 
     /// ## Example Usage
-    /// 
-    /// Basic usage:
+    /// ### Basic usage
     /// 
     /// ```csharp
     /// using Pulumi;
@@ -37,9 +36,48 @@ namespace Pulumi.Aws.Sagemaker
     /// 
     /// }
     /// ```
+    /// ### Code repository usage
+    /// 
+    /// ```csharp
+    /// using Pulumi;
+    /// using Aws = Pulumi.Aws;
+    /// 
+    /// class MyStack : Stack
+    /// {
+    ///     public MyStack()
+    ///     {
+    ///         var example = new Aws.Sagemaker.CodeRepository("example", new Aws.Sagemaker.CodeRepositoryArgs
+    ///         {
+    ///             CodeRepositoryName = "my-notebook-instance-code-repo",
+    ///             GitConfig = new Aws.Sagemaker.Inputs.CodeRepositoryGitConfigArgs
+    ///             {
+    ///                 RepositoryUrl = "https://github.com/terraform-providers/terraform-provider-aws.git",
+    ///             },
+    ///         });
+    ///         var ni = new Aws.Sagemaker.NotebookInstance("ni", new Aws.Sagemaker.NotebookInstanceArgs
+    ///         {
+    ///             RoleArn = aws_iam_role.Role.Arn,
+    ///             InstanceType = "ml.t2.medium",
+    ///             DefaultCodeRepository = example.CodeRepositoryName,
+    ///             Tags = 
+    ///             {
+    ///                 { "Name", "foo" },
+    ///             },
+    ///         });
+    ///     }
+    /// 
+    /// }
+    /// ```
     /// </summary>
     public partial class NotebookInstance : Pulumi.CustomResource
     {
+        /// <summary>
+        /// An array of up to three Git repositories to associate with the notebook instance.
+        /// These can be either the names of Git repositories stored as resources in your account, or the URL of Git repositories in [AWS CodeCommit](https://docs.aws.amazon.com/codecommit/latest/userguide/welcome.html) or in any other Git repository. These repositories are cloned at the same level as the default repository of your notebook instance.
+        /// </summary>
+        [Output("additionalCodeRepositories")]
+        public Output<ImmutableArray<string>> AdditionalCodeRepositories { get; private set; } = null!;
+
         /// <summary>
         /// The Amazon Resource Name (ARN) assigned by AWS to this notebook instance.
         /// </summary>
@@ -47,7 +85,7 @@ namespace Pulumi.Aws.Sagemaker
         public Output<string> Arn { get; private set; } = null!;
 
         /// <summary>
-        /// The Git repository associated with the notebook instance as its default code repository
+        /// The Git repository associated with the notebook instance as its default code repository. This can be either the name of a Git repository stored as a resource in your account, or the URL of a Git repository in [AWS CodeCommit](https://docs.aws.amazon.com/codecommit/latest/userguide/welcome.html) or in any other Git repository.
         /// </summary>
         [Output("defaultCodeRepository")]
         public Output<string?> DefaultCodeRepository { get; private set; } = null!;
@@ -83,6 +121,12 @@ namespace Pulumi.Aws.Sagemaker
         public Output<string> Name { get; private set; } = null!;
 
         /// <summary>
+        /// The network interface ID that Amazon SageMaker created at the time of creating the instance. Only available when setting `subnet_id`.
+        /// </summary>
+        [Output("networkInterfaceId")]
+        public Output<string> NetworkInterfaceId { get; private set; } = null!;
+
+        /// <summary>
         /// The ARN of the IAM role to be used by the notebook instance which allows SageMaker to call other services on your behalf.
         /// </summary>
         [Output("roleArn")]
@@ -111,6 +155,12 @@ namespace Pulumi.Aws.Sagemaker
         /// </summary>
         [Output("tags")]
         public Output<ImmutableDictionary<string, string>?> Tags { get; private set; } = null!;
+
+        /// <summary>
+        /// The URL that you use to connect to the Jupyter notebook that is running in your notebook instance.
+        /// </summary>
+        [Output("url")]
+        public Output<string> Url { get; private set; } = null!;
 
         /// <summary>
         /// The size, in GB, of the ML storage volume to attach to the notebook instance. The default value is 5 GB.
@@ -164,8 +214,21 @@ namespace Pulumi.Aws.Sagemaker
 
     public sealed class NotebookInstanceArgs : Pulumi.ResourceArgs
     {
+        [Input("additionalCodeRepositories")]
+        private InputList<string>? _additionalCodeRepositories;
+
         /// <summary>
-        /// The Git repository associated with the notebook instance as its default code repository
+        /// An array of up to three Git repositories to associate with the notebook instance.
+        /// These can be either the names of Git repositories stored as resources in your account, or the URL of Git repositories in [AWS CodeCommit](https://docs.aws.amazon.com/codecommit/latest/userguide/welcome.html) or in any other Git repository. These repositories are cloned at the same level as the default repository of your notebook instance.
+        /// </summary>
+        public InputList<string> AdditionalCodeRepositories
+        {
+            get => _additionalCodeRepositories ?? (_additionalCodeRepositories = new InputList<string>());
+            set => _additionalCodeRepositories = value;
+        }
+
+        /// <summary>
+        /// The Git repository associated with the notebook instance as its default code repository. This can be either the name of a Git repository stored as a resource in your account, or the URL of a Git repository in [AWS CodeCommit](https://docs.aws.amazon.com/codecommit/latest/userguide/welcome.html) or in any other Git repository.
         /// </summary>
         [Input("defaultCodeRepository")]
         public Input<string>? DefaultCodeRepository { get; set; }
@@ -255,6 +318,19 @@ namespace Pulumi.Aws.Sagemaker
 
     public sealed class NotebookInstanceState : Pulumi.ResourceArgs
     {
+        [Input("additionalCodeRepositories")]
+        private InputList<string>? _additionalCodeRepositories;
+
+        /// <summary>
+        /// An array of up to three Git repositories to associate with the notebook instance.
+        /// These can be either the names of Git repositories stored as resources in your account, or the URL of Git repositories in [AWS CodeCommit](https://docs.aws.amazon.com/codecommit/latest/userguide/welcome.html) or in any other Git repository. These repositories are cloned at the same level as the default repository of your notebook instance.
+        /// </summary>
+        public InputList<string> AdditionalCodeRepositories
+        {
+            get => _additionalCodeRepositories ?? (_additionalCodeRepositories = new InputList<string>());
+            set => _additionalCodeRepositories = value;
+        }
+
         /// <summary>
         /// The Amazon Resource Name (ARN) assigned by AWS to this notebook instance.
         /// </summary>
@@ -262,7 +338,7 @@ namespace Pulumi.Aws.Sagemaker
         public Input<string>? Arn { get; set; }
 
         /// <summary>
-        /// The Git repository associated with the notebook instance as its default code repository
+        /// The Git repository associated with the notebook instance as its default code repository. This can be either the name of a Git repository stored as a resource in your account, or the URL of a Git repository in [AWS CodeCommit](https://docs.aws.amazon.com/codecommit/latest/userguide/welcome.html) or in any other Git repository.
         /// </summary>
         [Input("defaultCodeRepository")]
         public Input<string>? DefaultCodeRepository { get; set; }
@@ -296,6 +372,12 @@ namespace Pulumi.Aws.Sagemaker
         /// </summary>
         [Input("name")]
         public Input<string>? Name { get; set; }
+
+        /// <summary>
+        /// The network interface ID that Amazon SageMaker created at the time of creating the instance. Only available when setting `subnet_id`.
+        /// </summary>
+        [Input("networkInterfaceId")]
+        public Input<string>? NetworkInterfaceId { get; set; }
 
         /// <summary>
         /// The ARN of the IAM role to be used by the notebook instance which allows SageMaker to call other services on your behalf.
@@ -338,6 +420,12 @@ namespace Pulumi.Aws.Sagemaker
             get => _tags ?? (_tags = new InputMap<string>());
             set => _tags = value;
         }
+
+        /// <summary>
+        /// The URL that you use to connect to the Jupyter notebook that is running in your notebook instance.
+        /// </summary>
+        [Input("url")]
+        public Input<string>? Url { get; set; }
 
         /// <summary>
         /// The size, in GB, of the ML storage volume to attach to the notebook instance. The default value is 5 GB.