+++ sdk/go/aws/cloudwatch/eventTarget.go
@@ -10,7 +10,9 @@ import (
 	"github.com/pulumi/pulumi/sdk/v2/go/pulumi"
 )
 
-// Provides a CloudWatch Event Target resource.
+// Provides an EventBridge Target resource.
+//
+// > **Note:** EventBridge was formerly known as CloudWatch Events. The functionality is identical.
 //
 // ## Example Usage
 //