+++ sdk/nodejs/glue/index.ts
@@ -12,6 +12,7 @@ export * from "./getScript";
 export * from "./job";
 export * from "./mltransform";
 export * from "./partition";
+export * from "./resourcePolicy";
 export * from "./securityConfiguration";
 export * from "./trigger";
 export * from "./userDefinedFunction";