+++ sdk/nodejs/s3/bucket.ts
@@ -13,6 +13,8 @@ import {CannedAcl, RoutingRule} from "./index";
 /**
  * Provides a S3 bucket resource.
  *
+ * > This functionality is for managing S3 in an AWS Partition. To manage [S3 on Outposts](https://docs.aws.amazon.com/AmazonS3/latest/dev/S3onOutposts.html), see the [`aws.s3control.Bucket` resource](https://www.terraform.io/docs/providers/aws/r/s3control_bucket.html).
+ *
  * ## Example Usage
  * ### Private Bucket w/ Tags
  *