+++ sdk/nodejs/sagemaker/notebookInstance.ts
@@ -8,8 +8,7 @@ import * as utilities from "../utilities";
  * Provides a Sagemaker Notebook Instance resource.
  *
  * ## Example Usage
- *
- * Basic usage:
+ * ### Basic usage
  *
  * ```typescript
  * import * as pulumi from "@pulumi/pulumi";
@@ -18,6 +17,27 @@ import * as utilities from "../utilities";
  * const ni = new aws.sagemaker.NotebookInstance("ni", {
  *     roleArn: aws_iam_role.role.arn,
  *     instanceType: "ml.t2.medium",
+ *     tags: {
+ *         Name: "foo",
+ *     },
+ * });
+ * ```
+ * ### Code repository usage
+ *
+ * ```typescript
+ * import * as pulumi from "@pulumi/pulumi";
+ * import * as aws from "@pulumi/aws";
+ *
+ * const example = new aws.sagemaker.CodeRepository("example", {
+ *     codeRepositoryName: "my-notebook-instance-code-repo",
+ *     gitConfig: {
+ *         repositoryUrl: "https://github.com/terraform-providers/terraform-provider-aws.git",
+ *     },
+ * });
+ * const ni = new aws.sagemaker.NotebookInstance("ni", {
+ *     roleArn: aws_iam_role.role.arn,
+ *     instanceType: "ml.t2.medium",
+ *     defaultCodeRepository: example.codeRepositoryName,
  *     tags: {
  *         Name: "foo",
  *     },
@@ -53,11 +73,16 @@ export class NotebookInstance extends pulumi.CustomResource {
     }
 
     /**
+     * An array of up to three Git repositories to associate with the notebook instance.
+     * These can be either the names of Git repositories stored as resources in your account, or the URL of Git repositories in [AWS CodeCommit](https://docs.aws.amazon.com/codecommit/latest/userguide/welcome.html) or in any other Git repository. These repositories are cloned at the same level as the default repository of your notebook instance.
+     */
+    public readonly additionalCodeRepositories!: pulumi.Output<string[] | undefined>;
+    /**
      * The Amazon Resource Name (ARN) assigned by AWS to this notebook instance.
      */
     public /*out*/ readonly arn!: pulumi.Output<string>;
     /**
-     * The Git repository associated with the notebook instance as its default code repository
+     * The Git repository associated with the notebook instance as its default code repository. This can be either the name of a Git repository stored as a resource in your account, or the URL of a Git repository in [AWS CodeCommit](https://docs.aws.amazon.com/codecommit/latest/userguide/welcome.html) or in any other Git repository.
      */
     public readonly defaultCodeRepository!: pulumi.Output<string | undefined>;
     /**
@@ -81,6 +106,10 @@ export class NotebookInstance extends pulumi.CustomResource {
      */
     public readonly name!: pulumi.Output<string>;
     /**
+     * The network interface ID that Amazon SageMaker created at the time of creating the instance. Only available when setting `subnetId`.
+     */
+    public /*out*/ readonly networkInterfaceId!: pulumi.Output<string>;
+    /**
      * The ARN of the IAM role to be used by the notebook instance which allows SageMaker to call other services on your behalf.
      */
     public readonly roleArn!: pulumi.Output<string>;
@@ -101,6 +130,10 @@ export class NotebookInstance extends pulumi.CustomResource {
      */
     public readonly tags!: pulumi.Output<{[key: string]: string} | undefined>;
     /**
+     * The URL that you use to connect to the Jupyter notebook that is running in your notebook instance.
+     */
+    public /*out*/ readonly url!: pulumi.Output<string>;
+    /**
      * The size, in GB, of the ML storage volume to attach to the notebook instance. The default value is 5 GB.
      */
     public readonly volumeSize!: pulumi.Output<number | undefined>;
@@ -117,6 +150,7 @@ export class NotebookInstance extends pulumi.CustomResource {
         let inputs: pulumi.Inputs = {};
         if (opts && opts.id) {
             const state = argsOrState as NotebookInstanceState | undefined;
+            inputs["additionalCodeRepositories"] = state ? state.additionalCodeRepositories : undefined;
             inputs["arn"] = state ? state.arn : undefined;
             inputs["defaultCodeRepository"] = state ? state.defaultCodeRepository : undefined;
             inputs["directInternetAccess"] = state ? state.directInternetAccess : undefined;
@@ -124,11 +158,13 @@ export class NotebookInstance extends pulumi.CustomResource {
             inputs["kmsKeyId"] = state ? state.kmsKeyId : undefined;
             inputs["lifecycleConfigName"] = state ? state.lifecycleConfigName : undefined;
             inputs["name"] = state ? state.name : undefined;
+            inputs["networkInterfaceId"] = state ? state.networkInterfaceId : undefined;
             inputs["roleArn"] = state ? state.roleArn : undefined;
             inputs["rootAccess"] = state ? state.rootAccess : undefined;
             inputs["securityGroups"] = state ? state.securityGroups : undefined;
             inputs["subnetId"] = state ? state.subnetId : undefined;
             inputs["tags"] = state ? state.tags : undefined;
+            inputs["url"] = state ? state.url : undefined;
             inputs["volumeSize"] = state ? state.volumeSize : undefined;
         } else {
             const args = argsOrState as NotebookInstanceArgs | undefined;
@@ -138,6 +174,7 @@ export class NotebookInstance extends pulumi.CustomResource {
             if (!args || args.roleArn === undefined) {
                 throw new Error("Missing required property 'roleArn'");
             }
+            inputs["additionalCodeRepositories"] = args ? args.additionalCodeRepositories : undefined;
             inputs["defaultCodeRepository"] = args ? args.defaultCodeRepository : undefined;
             inputs["directInternetAccess"] = args ? args.directInternetAccess : undefined;
             inputs["instanceType"] = args ? args.instanceType : undefined;
@@ -151,6 +188,8 @@ export class NotebookInstance extends pulumi.CustomResource {
             inputs["tags"] = args ? args.tags : undefined;
             inputs["volumeSize"] = args ? args.volumeSize : undefined;
             inputs["arn"] = undefined /*out*/;
+            inputs["networkInterfaceId"] = undefined /*out*/;
+            inputs["url"] = undefined /*out*/;
         }
         if (!opts) {
             opts = {}
@@ -168,11 +207,16 @@ export class NotebookInstance extends pulumi.CustomResource {
  */
 export interface NotebookInstanceState {
     /**
+     * An array of up to three Git repositories to associate with the notebook instance.
+     * These can be either the names of Git repositories stored as resources in your account, or the URL of Git repositories in [AWS CodeCommit](https://docs.aws.amazon.com/codecommit/latest/userguide/welcome.html) or in any other Git repository. These repositories are cloned at the same level as the default repository of your notebook instance.
+     */
+    readonly additionalCodeRepositories?: pulumi.Input<pulumi.Input<string>[]>;
+    /**
      * The Amazon Resource Name (ARN) assigned by AWS to this notebook instance.
      */
     readonly arn?: pulumi.Input<string>;
     /**
-     * The Git repository associated with the notebook instance as its default code repository
+     * The Git repository associated with the notebook instance as its default code repository. This can be either the name of a Git repository stored as a resource in your account, or the URL of a Git repository in [AWS CodeCommit](https://docs.aws.amazon.com/codecommit/latest/userguide/welcome.html) or in any other Git repository.
      */
     readonly defaultCodeRepository?: pulumi.Input<string>;
     /**
@@ -196,6 +240,10 @@ export interface NotebookInstanceState {
      */
     readonly name?: pulumi.Input<string>;
     /**
+     * The network interface ID that Amazon SageMaker created at the time of creating the instance. Only available when setting `subnetId`.
+     */
+    readonly networkInterfaceId?: pulumi.Input<string>;
+    /**
      * The ARN of the IAM role to be used by the notebook instance which allows SageMaker to call other services on your behalf.
      */
     readonly roleArn?: pulumi.Input<string>;
@@ -216,6 +264,10 @@ export interface NotebookInstanceState {
      */
     readonly tags?: pulumi.Input<{[key: string]: pulumi.Input<string>}>;
     /**
+     * The URL that you use to connect to the Jupyter notebook that is running in your notebook instance.
+     */
+    readonly url?: pulumi.Input<string>;
+    /**
      * The size, in GB, of the ML storage volume to attach to the notebook instance. The default value is 5 GB.
      */
     readonly volumeSize?: pulumi.Input<number>;
@@ -226,7 +278,12 @@ export interface NotebookInstanceState {
  */
 export interface NotebookInstanceArgs {
     /**
-     * The Git repository associated with the notebook instance as its default code repository
+     * An array of up to three Git repositories to associate with the notebook instance.
+     * These can be either the names of Git repositories stored as resources in your account, or the URL of Git repositories in [AWS CodeCommit](https://docs.aws.amazon.com/codecommit/latest/userguide/welcome.html) or in any other Git repository. These repositories are cloned at the same level as the default repository of your notebook instance.
+     */
+    readonly additionalCodeRepositories?: pulumi.Input<pulumi.Input<string>[]>;
+    /**
+     * The Git repository associated with the notebook instance as its default code repository. This can be either the name of a Git repository stored as a resource in your account, or the URL of a Git repository in [AWS CodeCommit](https://docs.aws.amazon.com/codecommit/latest/userguide/welcome.html) or in any other Git repository.
      */
     readonly defaultCodeRepository?: pulumi.Input<string>;
     /**