+++ sdk/python/pulumi_aws/cloudwatch/event_permission.py
@@ -25,7 +25,9 @@ class EventPermission(pulumi.CustomResource):
                  __name__=None,
                  __opts__=None):
         """
-        Provides a resource to create a CloudWatch Events permission to support cross-account events in the current account default event bus.
+        Provides a resource to create an EventBridge permission to support cross-account events in the current account default event bus.
+
+        > **Note:** EventBridge was formerly known as CloudWatch Events. The functionality is identical.
 
         ## Example Usage
         ### Account Access