+++ sdk/python/pulumi_aws/s3/access_point.py
@@ -30,7 +30,7 @@ class AccessPoint(pulumi.CustomResource):
         Provides a resource to manage an S3 Access Point.
 
         ## Example Usage
-        ### Basic Usage
+        ### AWS Partition Bucket
 
         ```python
         import pulumi
@@ -39,16 +39,16 @@ class AccessPoint(pulumi.CustomResource):
         example_bucket = aws.s3.Bucket("exampleBucket")
         example_access_point = aws.s3.AccessPoint("exampleAccessPoint", bucket=example_bucket.id)
         ```
-        ### Access Point Restricted to a VPC
+        ### S3 on Outposts Bucket
 
         ```python
         import pulumi
         import pulumi_aws as aws
 
-        example_bucket = aws.s3.Bucket("exampleBucket")
+        example_bucket = aws.s3control.Bucket("exampleBucket", bucket="example")
         example_vpc = aws.ec2.Vpc("exampleVpc", cidr_block="10.0.0.0/16")
         example_access_point = aws.s3.AccessPoint("exampleAccessPoint",
-            bucket=example_bucket.id,
+            bucket=example_bucket.arn,
             vpc_configuration=aws.s3.AccessPointVpcConfigurationArgs(
                 vpc_id=example_vpc.id,
             ))
@@ -57,11 +57,11 @@ class AccessPoint(pulumi.CustomResource):
         :param str resource_name: The name of the resource.
         :param pulumi.ResourceOptions opts: Options for the resource.
         :param pulumi.Input[str] account_id: The AWS account ID for the owner of the bucket for which you want to create an access point. Defaults to automatically determined account ID of the provider.
-        :param pulumi.Input[str] bucket: The name of the bucket that you want to associate this access point with.
+        :param pulumi.Input[str] bucket: The name of an AWS Partition S3 Bucket or the Amazon Resource Name (ARN) of S3 on Outposts Bucket that you want to associate this access point with.
         :param pulumi.Input[str] name: The name you want to assign to this access point.
         :param pulumi.Input[str] policy: A valid JSON document that specifies the policy that you want to apply to this access point.
         :param pulumi.Input[pulumi.InputType['AccessPointPublicAccessBlockConfigurationArgs']] public_access_block_configuration: Configuration block to manage the `PublicAccessBlock` configuration that you want to apply to this Amazon S3 bucket. You can enable the configuration options in any combination. Detailed below.
-        :param pulumi.Input[pulumi.InputType['AccessPointVpcConfigurationArgs']] vpc_configuration: Configuration block to restrict access to this access point to requests from the specified Virtual Private Cloud (VPC). Detailed below.
+        :param pulumi.Input[pulumi.InputType['AccessPointVpcConfigurationArgs']] vpc_configuration: Configuration block to restrict access to this access point to requests from the specified Virtual Private Cloud (VPC). Required for S3 on Outposts. Detailed below.
         """
         if __name__ is not None:
             warnings.warn("explicit use of __name__ is deprecated", DeprecationWarning)
@@ -121,7 +121,7 @@ class AccessPoint(pulumi.CustomResource):
         :param pulumi.ResourceOptions opts: Options for the resource.
         :param pulumi.Input[str] account_id: The AWS account ID for the owner of the bucket for which you want to create an access point. Defaults to automatically determined account ID of the provider.
         :param pulumi.Input[str] arn: Amazon Resource Name (ARN) of the S3 Access Point.
-        :param pulumi.Input[str] bucket: The name of the bucket that you want to associate this access point with.
+        :param pulumi.Input[str] bucket: The name of an AWS Partition S3 Bucket or the Amazon Resource Name (ARN) of S3 on Outposts Bucket that you want to associate this access point with.
         :param pulumi.Input[str] domain_name: The DNS domain name of the S3 Access Point in the format _`name`_-_`account_id`_.s3-accesspoint._region_.amazonaws.com.
                Note: S3 access points only support secure access by HTTPS. HTTP isn't supported.
         :param pulumi.Input[bool] has_public_access_policy: Indicates whether this access point currently has a policy that allows public access.
@@ -129,7 +129,7 @@ class AccessPoint(pulumi.CustomResource):
         :param pulumi.Input[str] network_origin: Indicates whether this access point allows access from the public Internet. Values are `VPC` (the access point doesn't allow access from the public Internet) and `Internet` (the access point allows access from the public Internet, subject to the access point and bucket access policies).
         :param pulumi.Input[str] policy: A valid JSON document that specifies the policy that you want to apply to this access point.
         :param pulumi.Input[pulumi.InputType['AccessPointPublicAccessBlockConfigurationArgs']] public_access_block_configuration: Configuration block to manage the `PublicAccessBlock` configuration that you want to apply to this Amazon S3 bucket. You can enable the configuration options in any combination. Detailed below.
-        :param pulumi.Input[pulumi.InputType['AccessPointVpcConfigurationArgs']] vpc_configuration: Configuration block to restrict access to this access point to requests from the specified Virtual Private Cloud (VPC). Detailed below.
+        :param pulumi.Input[pulumi.InputType['AccessPointVpcConfigurationArgs']] vpc_configuration: Configuration block to restrict access to this access point to requests from the specified Virtual Private Cloud (VPC). Required for S3 on Outposts. Detailed below.
         """
         opts = pulumi.ResourceOptions.merge(opts, pulumi.ResourceOptions(id=id))
 
@@ -167,7 +167,7 @@ class AccessPoint(pulumi.CustomResource):
     @pulumi.getter
     def bucket(self) -> pulumi.Output[str]:
         """
-        The name of the bucket that you want to associate this access point with.
+        The name of an AWS Partition S3 Bucket or the Amazon Resource Name (ARN) of S3 on Outposts Bucket that you want to associate this access point with.
         """
         return pulumi.get(self, "bucket")
 
@@ -224,7 +224,7 @@ class AccessPoint(pulumi.CustomResource):
     @pulumi.getter(name="vpcConfiguration")
     def vpc_configuration(self) -> pulumi.Output[Optional['outputs.AccessPointVpcConfiguration']]:
         """
-        Configuration block to restrict access to this access point to requests from the specified Virtual Private Cloud (VPC). Detailed below.
+        Configuration block to restrict access to this access point to requests from the specified Virtual Private Cloud (VPC). Required for S3 on Outposts. Detailed below.
         """
         return pulumi.get(self, "vpc_configuration")